backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1
pydantic[email]==2.7.2
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
celery==5.4.0
redis==5.0.4
psycopg2-binary==2.9.9
cryptography==42.0.7
httpx==0.27.0
python-dotenv==1.0.1
boto3==1.34.116
stripe==9.8.0
anthropic==0.28.0
jinja2==3.1.4
python-magic==0.4.27
pyopenssl==24.1.0
pytest==8.2.1
pytest-asyncio==0.23.7
pyotp==2.9.0
